yuyusrequirement.txt
cmake
opencv-python
face-recognition
numpy
fastapi[all]
sqlalchemy
alembic
psycopg2
asyncpg
pycryptodome
bcrypt
fastapi-sessions
uvicorn
python-multipart
user-agents
python-dotenv
pyjwt
passlib
python-jose[cryptography]
pillow
pytz
ntplib
requests
tzdata